3d arrows & scaling

+++ src/glyph2drendermodel/setglyphdim.cc
@@ -5,7 +5,7 @@ void Glyph2dRenderModel::set_glyph_dim(size_t dim)
     d_glyph_dim = dim;
 
     float d_coord_advance = 
-        static_cast<float>(250) / static_cast<float>(d_glyph_dim);
+        250.0f / static_cast<float>(d_glyph_dim);
     float d_coord_start = d_coord_advance / 2.0f;
 
     /* Upload glyph positions once */
